multipart: Support parameters on file Content-Type
This commit adds support for passing the full contents of the Content-
Type header for a file event, in order that a second instance of BusBoy
can be used to read nested multipart content. This is common when used
for parsing some EDI protocols - for example:

```
Content-Type: multipart/form-data; boundary="boundaryKU3VUg=="

--boundaryKU3VUg==
Content-Disposition: form-data; name="from"

A Sender
--boundaryKU3VUg==
Content-Disposition: form-data; name="input-data"; filename="edi.dat"
Content-Type: multipart/encrypted;
 protocol="application/pgp-encrypted"; boundary="boundary2AW02w=="

// More multipart content using the new boundary omitted here
```

So as not to change the existing behaviour, the full content type header
is presented as an additional argument to the `file` event. This
resulted in some changes to existing tests, since the optional final
argument will be the same as the fifth if no parameters are present in
the Content-Type header of the part.