Update engines field (#91)
* Update engines field

In #89 support for Node.js 10 and 12 were dropped. It didn't explicitly mention a minimum version, but seems to be testing on the latest 14.x. I chose 14.13.1 since that is the minimum version of Node.js 14.x that Multer supports, and I would be very happy if this module would also support it.

* Relax engines field to Node.js 14.0.0

Co-authored-by: Uzlopak <aras.abbasi@googlemail.com>

Co-authored-by: Uzlopak <aras.abbasi@googlemail.com>